requirements.txt
pymongo
hikari
hikari-crescent
hikari-toolbox
asyncio
html2image
motor